NutriNow_BackEnd/requirements.txt
mysql-connector-python
python-dotenv
langchain
langchain-community
langchain-core
pydantic
Flask
Flask-Cors
Werkzeug
ollama
requests
oauthlib